recreate infra_core, and move select files (model_handlers, constants, models) into it. Update references. All tests pass.
