+++ infra_core/model_handlers/__init__.py
@@ -1,0 +1,11 @@
+"""
+Model Handlers Package
+
+Provides interfaces and implementations for different model backends.
+"""
+
+from infra_core.model_handlers.base import BaseModelHandler
+from infra_core.model_handlers.openai_handler import OpenAIModelHandler
+from infra_core.model_handlers.ollama_handler import OllamaModelHandler
+
+__all__ = ["BaseModelHandler", "OpenAIModelHandler", "OllamaModelHandler"]
