+++ legacy_logseq/tools/broadcast_queue_tool.py
@@ -6,7 +6,7 @@ from autogen_core.tools import FunctionTool
 from pathlib import Path
 
 from legacy_logseq.memory.memory_bank import FileMemoryBank
-from legacy_logseq.constants import (
+from infra_core.constants import (
     MEMORY_BANKS_ROOT,
     BROADCAST_QUEUE_PROJECT,
     BROADCAST_QUEUE_SESSION,
